requirements.txt
Django==3.2.6
djangorestframework==3.12.4
Pillow==10.4.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
django-import-export==4.1.1
django-nested-admin==4.1.1
drf-spectacular==0.27.2
openai==1.51.0
pyjwt==2.9.0
django-environ==0.11.2
django-silk==5.1.0
django-cors-headers==4.2.0
django-storages==1.14.4
boto3==1.35.37
gunicorn==23.0.0 # need to update the Procfile for gunicorn
djangorestframework-xml==2.0.0
django-silk==5.1.0
google-auth==2.35.0
google-cloud==0.34.0
google-cloud-secret-manager==2.7.2
# celery[redis]==5.2.7